Move B1500Module to its own module KeysightB1500_module.py
